+++ CHANGELOG.md
@@ -313,4 +313,17 @@
   suggested in the [documentation](https://pytorch.org/tutorials//intermediate/torch_compile_tutorial.html). It
   increases the performance by about 1/100.
 
+## 0.9.7
+
+- New models release with more meta-data.
+- Add a feature to use an AddressParser from a URI.
+- Add a feature to upload the trained model to a URI.
+- Add an example of how to use URI for parsing from and uploading to.
+- Improve error handling of `path_to_retrain_model`.
+- Bug-fix pre-processor error.
+- Add verbose override and improve verbosity handling in retrain.
+- Bug-fix the broken FastText installation using `fasttext-wheel` instead of `fasttext` (
+  see [here](https://github.com/facebookresearch/fastText/issues/512#issuecomment-1534519551)
+  and [here](https://github.com/facebookresearch/fastText/pull/1292)).
+
 ## dev